@@ -14,35 +14,57 @@ if TYPE_CHECKING:
 
 
 class Dataset2D:
-    param_names = ("store_type", "chunks")
+    param_names = ("store_type", "chunks", "array_type")
     params = (
         ("zarr", "h5ad"),
         ((-1,), None),
+        ("cat", "numeric", "string-array", "nullable-string-array"),
     )
 
     def setup_cache(self):
-        n_obs = 100000
-        df = pd.DataFrame(
-            {
-                "a": pd.Categorical(np.array(["a"] * n_obs)),
-                "b": np.arange(n_obs),
-            },
-            index=[f"cell{i}" for i in range(n_obs)],
-        )
-        for store in [
-            h5py.File("data.h5ad", mode="w"),
-            zarr.open("data.zarr", mode="w", zarr_version=2),
-        ]:
-            ad.io.write_elem(store, "obs", df)
+        n_obs = 10000
+        array_types = {
+            "numeric": np.arange(n_obs),
+            "string-array": np.array(["a"] * n_obs),
+            "nullable-string-array": pd.array(
+                ["a", pd.NA] * (n_obs // 2), dtype="string"
+            ),
+            "cat": pd.Categorical(np.array(["a"] * n_obs)),
+        }
+        for k, v in array_types.items():
+            for store in [
+                h5py.File(f"data_{k}.h5ad", mode="w"),
+                zarr.open(f"data_{k}.zarr", mode="w", zarr_version=2),
+            ]:
+                df = pd.DataFrame({"a": v}, index=[f"cell{i}" for i in range(n_obs)])
+                if writing_string_array_on_disk := (
+                    isinstance(v, np.ndarray) and df["a"].dtype == "string"
+                ):
+                    df["a"] = df["a"].to_numpy()
+                with ad.settings.override(allow_write_nullable_strings=True):
+                    ad.io.write_elem(store, "df", df)
+                if writing_string_array_on_disk:
+                    assert store["df"]["a"].attrs["encoding-type"] == "string-array"
 
-    def setup(self, store_type: Literal["zarr", "h5ad"], chunks: None | tuple[int]):
-        store = (
-            h5py.File("data.h5ad", mode="r")
+    def setup(
+        self,
+        store_type: Literal["zarr", "h5ad"],
+        chunks: None | tuple[int],
+        array_type: Literal["cat", "numeric", "string-array", "nullable-string-array"],
+    ):
+        self.store = (
+            h5py.File(f"data_{array_type}.h5ad", mode="r")
             if store_type == "h5ad"
-            else zarr.open("data.zarr")
+            else zarr.open(f"data_{array_type}.zarr")
         )
-        self.ds = ad.experimental.read_elem_lazy(store["obs"], chunks=chunks)
+        self.ds = ad.experimental.read_elem_lazy(self.store["df"], chunks=chunks)
         self.n_obs = self.ds.shape[0]
+
+    def time_read_lazy_default(self, *_):
+        ad.experimental.read_elem_lazy(self.store["df"])
+
+    def peakmem_read_lazy_default(self, *_):
+        ad.experimental.read_elem_lazy(self.store["df"])
 
     def time_getitem_slice(self, *_):
         self.ds.iloc[0 : (self.n_obs // 2)].to_memory()
@@ -61,3 +83,7 @@ class Dataset2D:
 
     def peakmem_getitem_bool_mask(self, *_):
         self.ds.iloc[np.random.randint(0, self.n_obs, self.n_obs // 2)].to_memory()
+
+    def time_concat(self, *_):
+        adatas = [ad.AnnData(obs=self.ds)] * 50
+        ad.concat(adatas, join="outer")
