@@ -15,6 +15,7 @@ import h5py
 import numpy as np
 import pandas as pd
 import pytest
+import zarr
 from packaging.version import Version
 from pandas.api.types import is_numeric_dtype
 from scipy import sparse
@@ -36,6 +37,7 @@ from anndata.compat import (
     XDataArray,
     XDataset,
     ZarrArray,
+    ZarrGroup,
     is_zarr_v2,
 )
 from anndata.utils import asarray
@@ -43,7 +45,6 @@ from anndata.utils import asarray
 if TYPE_CHECKING:
     from collections.abc import Callable, Collection, Iterable
     from typing import Any, Literal, TypeGuard, TypeVar
-
     from numpy.typing import NDArray
     from zarr.abc.store import ByteRequest
     from zarr.core.buffer import BufferPrototype
@@ -51,8 +52,7 @@ if TYPE_CHECKING:
     from .._types import ArrayStorageType
     from ..compat import Index1D
 
-    DT = TypeVar("DT")
-    _SubsetFunc = Callable[[pd.Index[str], int], Index1D]
+    type _SubsetFunc = Callable[[pd.Index[str], int], Index1D]
 
 
 try:
@@ -135,7 +135,7 @@ def gen_vstr_recarray(m, n, dtype=None):
     )
 
 
-def issubdtype(
+def issubdtype[DT](
     a: np.dtype | pd.api.extensions.ExtensionDtype | type,
     b: type[DT] | tuple[type[DT], ...],
 ) -> TypeGuard[DT]:
@@ -1366,3 +1366,23 @@ def get_multiindex_columns_df(shape: tuple[int, int]) -> pd.DataFrame:
             + list(itertools.product(["b"], range(shape[1] // 2)))
         ),
     )
+
+
+def visititems_zarr(
+    z: ZarrGroup, visitor: Callable[[str, ZarrGroup | zarr.Array], None]
+) -> None:
+    for key in z:
+        maybe_group = z[key]
+        if isinstance(maybe_group, ZarrGroup):
+            visititems_zarr(maybe_group, visitor)
+        else:
+            visitor(key, maybe_group)
+
+
+def check_all_sharded(g: ZarrGroup):
+    def visit(key: str, arr: zarr.Array | zarr.Group):
+        # Check for recarray via https://numpy.org/doc/stable/user/basics.rec.html#manipulating-and-displaying-structured-datatypes
+        if isinstance(arr, zarr.Array) and arr.shape != () and arr.dtype.names is None:
+            assert arr.shards is not None
+
+    visititems_zarr(g, visitor=visit)
