@@ -72,7 +72,7 @@ def _alloc_cache():
 # As of 2025.09.* dask, this needs a bit more than the previous 1.5mb.
 # TODO: Why?
 @pytest.mark.usefixtures("_alloc_cache")
-@pytest.mark.limit_memory("1.7 MB")
+@pytest.mark.limit_memory("2.2 MB")
 def test_size_of_view(mapping_name, give_chunks):
     import dask.array as da
 
